LayoutTests/fast/events/message-port-expected.txt

CONSOLE MESSAGE: line 20: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Test cross-frame MessagePort communication.

Should be a series of SUCCESS messages, followed with DONE.
